Add files via upload
The cleaning of the data was carried out from raw MOSPI files of DCT and TXT type by Dr Jihye Kim and Dr Ioana Macoveciuc of the University of Manchester. Please be sure to acknowledge them and cite your source if you adapt this code to your own needs.  Thank you!  Wendy Olsen